+++ lib/prefix_tree.h
@@ -8,7 +8,7 @@
 
 typedef struct _PrefixTreeNode {
   struct _PrefixTreeNode **nextLetters;
-  unsigned int value;
+  unsigned long value;
 } PrefixTreeNode;
 
 typedef struct _PrefixTree {
@@ -16,7 +16,7 @@ typedef struct _PrefixTree {
 } PrefixTree;
 
 PrefixTree *prefix_tree_init();
-void prefix_tree_add_word(PrefixTree *tree, char *word, unsigned int value);
+void prefix_tree_add_word(PrefixTree *tree, char *word, unsigned long value);
 unsigned int prefix_tree_has_word(PrefixTree *tree, char *word);
 void prefix_tree_destroy(PrefixTree *tree);
